fix(scm): address PR #253 third-pass review notes
Move SCM test doubles out of production protocol.py, add
download_workflow_run_logs on GitHubClient/adapter, and remove
the ctx.github __getattr__/__setattr__ shim after migrating tests.
