+++ src/mergecraft/ci/providers/github_actions.py
@@ -12,7 +12,7 @@ from loguru import logger
 
 from mergecraft.ci.log_excerpt import analyze_log
 from mergecraft.ci.truncate import DEFAULT_TRUNCATION_CAP, apply_truncation
-from mergecraft.scm.github import github_client_from_scm
+from mergecraft.scm.github import GitHubScmAdapter
 
 if TYPE_CHECKING:
     from mergecraft.ci.types import ProviderContext, RawFailure
@@ -64,17 +64,14 @@ class GitHubActionsProvider:
         for run in selected:
             run_id = run["id"]
             try:
-                github = github_client_from_scm(ctx.scm)
-                if github is None:
+                scm = ctx.scm
+                if not isinstance(scm, GitHubScmAdapter):
                     raise RuntimeError("log download requires a GitHub SCM adapter")
-                response = await github._client.get(
-                    f"/repos/{ctx.repo.owner}/{ctx.repo.name}/actions/runs/{run_id}/logs",
-                    headers={"Accept": "application/vnd.github+json"},
-                    follow_redirects=True,
+                raw = await scm.download_workflow_run_logs(
+                    ctx.repo.owner,
+                    ctx.repo.name,
+                    run_id,
                 )
-                if response.status_code >= 400:
-                    raise RuntimeError(f"log download failed: {response.status_code}")
-                raw = response.content
             except Exception as err:
                 logger.info("failed to download logs for run {}: {}", run_id, err)
                 continue
